feat: sql primary key. util/db_promise.js
