+++ models/user.js
@@ -1,22 +1,7 @@
 //MVC: user model
-const sqlite3 = require("sqlite3").verbose();
-const db = new sqlite3.Database("message_db.db");
 
 const Error_Code = require("../util/error_code");
-
-//await wrapper for sqlite3
-//TODO: sqlite3 doesn't support Promise, can't use await. maybe use sqlite?
-async function db_all(query){
-	return new Promise(function(resolve, reject){
-		db.all(query, function(err, rows){
-			if(err){
-				reject(err);
-			}else{
-				resolve(rows);
-			}
-		});
-	});
-}
+const DB_Promise = require("../util/db_promise");
 
 class User{
 	constructor(username, password){
@@ -29,6 +14,16 @@ class User{
 
 		//TODO: use SQL primary key. only use SQL INSERT without SQL SELECT
 		//TODO: sqlite3 doesn't support Promise, can't use await. maybe use sqlite?
+		try{
+			await DB_Promise.db_run(`INSERT INTO accounts VALUES(\"${ this.username }\", \"${ this.password }\")`);
+		}catch(err){
+			console.log("[error] register(): " + err);
+			throw err;
+		}
+
+		console.log("register(): success.");
+
+		/*
 		let account_rows;
 		try{
 			account_rows = await db_all(`SELECT * FROM accounts WHERE username = \"${ this.username }\"`);
@@ -46,6 +41,7 @@ class User{
 			db.run("INSERT INTO accounts VALUES(?, ?)", this.username, this.password);
 			return;
 		}
+		*/
 	}
 
 	login(){
